+++ common/src/main/java/com/zigythebird/playeranim/animation/layered/modifier/AbstractFadeModifier.java
@@ -114,7 +114,7 @@ public abstract class AbstractFadeModifier extends AbstractModifier {
         return new AbstractFadeModifier(length) {
             @Override
             protected float getAlpha(PlayerAnimBone bone, float progress) {
-                return ease.buildTransformer(null).apply(progress);
+                return ease.buildTransformer(null).apply((double) progress).floatValue();
             }
         };
     }
